Remove old affiliate section
